Conversion fix.

+++ src/Game/GameScenes/DanielScene.cpp
@@ -163,7 +163,7 @@ void DanielScene::Update(float deltaTime) {
         postProcessing->ApplyFilter(fxaaFilter);
     }
     
-    gammaCorrectionFilter->SetBrightness(GameSettings::GetInstance().GetDouble("Gamma"));
+    gammaCorrectionFilter->SetBrightness(static_cast<float>(GameSettings::GetInstance().GetDouble("Gamma")));
     postProcessing->ApplyFilter(gammaCorrectionFilter);
     
     postProcessing->Render();
